requirements.txt
Pillow
PyQt5
PyQt5-Qt5
lxml
python-docx
setuptools